+++ libMx/d_mdrive_dio.c
@@ -294,7 +294,7 @@ mxd_mdrive_din_read( MX_DIGITAL_INPUT *dinput )
 			dinput->record->name );
 	}
 
-	sprintf( command, "PR I%d", mdrive_dinput->port_number );
+	sprintf( command, "PR I%d", (int) mdrive_dinput->port_number );
 
 	mx_status = mxd_mdrive_command( mdrive, command,
 				response, sizeof( response ),
@@ -406,7 +406,7 @@ mxd_mdrive_dout_write( MX_DIGITAL_OUTPUT *doutput )
 	}
 
 	sprintf( command, "I%d=%lu",
-			mdrive_doutput->port_number,
+			(int) mdrive_doutput->port_number,
 			(unsigned long) doutput->value );
 
 	mx_status = mxd_mdrive_command( mdrive, command,
